Lava elevation fix, stranded optimisation and SM shard removal (#3002)
* pve wave defense mode tweaks

* optimises spawn pick

* wave spawn cache

* no sm shard

* lava elevation fix

* whitespace

* firefight/stranded extra waves

* last_spawns_list tweaks + stranded mob spawn list modifications

+++ code/game/turfs/unsimulated/lava.dm
@@ -45,12 +45,13 @@
 		loseme = 1
 
 	if(loseme)
-		for(var/obj/effect/decal/cleanable/ash/A in src)
-			qdel(A)
-		spawn(rand(25,75))
-			if(Obj && Obj.loc == src)
-				src.visible_message("\icon[Obj]<span class='danger'>[Obj] sinks under the surface of [src]!</span>")
-				qdel(Obj)
+		if(Obj.elevation == BASE_ELEVATION)
+			for(var/obj/effect/decal/cleanable/ash/A in src)
+				qdel(A)
+			spawn(rand(25,75))
+				if(Obj && Obj.loc == src)
+					src.visible_message("\icon[Obj]<span class='danger'>[Obj] sinks under the surface of [src]!</span>")
+					qdel(Obj)
 
 	light_range = 5
 	light_power = 2
